dip(restructuring widget selection)
