+++ app/client/src/constants/ReduxActionConstants.tsx
@@ -290,6 +290,7 @@ export const ReduxActionTypes: { [key: string]: string } = {
   INVITED_USER_SIGNUP_SUCCESS: "INVITED_USER_SIGNUP_SUCCESS",
   INVITED_USER_SIGNUP_INIT: "INVITED_USER_SIGNUP_INIT",
   DISABLE_WIDGET_DRAG: "DISABLE_WIDGET_DRAG",
+  SELECT_WIDGET_INIT: "SELECT_WIDGET_INIT",
   SELECT_WIDGET: "SELECT_WIDGET",
   SELECT_MULTIPLE_WIDGETS: "SELECT_MULTIPLE_WIDGETS",
   SELECT_MULTIPLE_WIDGETS_INIT: "SELECT_MULTIPLE_WIDGETS_INIT",
