+++ app/client/src/pages/Editor/WidgetsEditor.tsx
@@ -11,7 +11,6 @@ import {
 import Centered from "components/designSystems/appsmith/CenteredWrapper";
 import EditorContextProvider from "components/editorComponents/EditorContextProvider";
 import { Spinner } from "@blueprintjs/core";
-import { useWidgetSelection } from "utils/hooks/dragResizeHooks";
 import AnalyticsUtil from "utils/AnalyticsUtil";
 import * as log from "loglevel";
 import { getCanvasClassName } from "utils/generators";
@@ -26,6 +25,7 @@ import { MainContainerLayoutControl } from "./MainContainerLayoutControl";
 import { useDynamicAppLayout } from "utils/hooks/useDynamicAppLayout";
 import Debugger from "components/editorComponents/Debugger";
 import { closePropertyPane } from "actions/widgetActions";
+import { useWidgetSelection } from "utils/hooks/useWidgetSelection";
 
 const EditorWrapper = styled.div`
   display: flex;
