+++ app/client/src/sagas/WidgetSelectionSagas.ts
@@ -1,0 +1,79 @@
+import { ReduxAction, ReduxActionTypes } from "constants/ReduxActionConstants";
+import { MAIN_CONTAINER_WIDGET_ID } from "constants/WidgetConstants";
+import { all, put, select, takeLatest } from "redux-saga/effects";
+import { getWidgetImmediateChildren, getWidgets } from "./selectors";
+import log from "loglevel";
+import {
+  selectAllWidgetsAction,
+  selectWidgetAction,
+} from "actions/widgetActions";
+import { Toaster } from "components/ads/Toast";
+import { createMessage, SELECT_ALL_WIDGETS_MSG } from "constants/messages";
+import { Variant } from "components/ads/common";
+
+// The following is computed to be used in the entity explorer
+// Every time a widget is selected, we need to expand widget entities
+// in the entity explorer so that the selected widget is visible
+function* selectedWidgetAncestrySaga(
+  action: ReduxAction<{ widgetId: string }>,
+) {
+  try {
+    const canvasWidgets = yield select(getWidgets);
+    const widgetIdsExpandList = [];
+    const selectedWidget = action.payload.widgetId;
+
+    // Make sure that the selected widget exists in canvasWidgets
+    let widgetId = canvasWidgets[selectedWidget]
+      ? canvasWidgets[selectedWidget].parentId
+      : undefined;
+    // If there is a parentId for the selectedWidget
+    if (widgetId) {
+      // Keep including the parent until we reach the main container
+      while (widgetId !== MAIN_CONTAINER_WIDGET_ID) {
+        widgetIdsExpandList.push(widgetId);
+        if (canvasWidgets[widgetId] && canvasWidgets[widgetId].parentId)
+          widgetId = canvasWidgets[widgetId].parentId;
+        else break;
+      }
+    }
+    yield put({
+      type: ReduxActionTypes.SET_SELECTED_WIDGET_ANCESTORY,
+      payload: widgetIdsExpandList,
+    });
+  } catch (error) {
+    log.debug("Could not compute selected widget's ancestry", error);
+  }
+}
+
+function* selectAllWidgetsSaga() {
+  const allWidgetsOnMainContainer: string[] = yield select(
+    getWidgetImmediateChildren,
+    MAIN_CONTAINER_WIDGET_ID,
+  );
+  if (allWidgetsOnMainContainer && allWidgetsOnMainContainer.length) {
+    yield put(selectAllWidgetsAction(allWidgetsOnMainContainer));
+    Toaster.show({
+      text: createMessage(SELECT_ALL_WIDGETS_MSG),
+      variant: Variant.info,
+      duration: 3000,
+    });
+  }
+}
+
+function* selectWidgetSaga(
+  action: ReduxAction<{ widgetId: string; isMultiSelect: boolean }>,
+) {
+  const { isMultiSelect, widgetId } = action.payload;
+  yield put(selectWidgetAction(widgetId, isMultiSelect));
+}
+
+export function* widgetSelectionSagas() {
+  yield all([
+    takeLatest(ReduxActionTypes.SELECT_WIDGET_INIT, selectedWidgetAncestrySaga),
+    takeLatest(ReduxActionTypes.SELECT_WIDGET_INIT, selectWidgetSaga),
+    takeLatest(
+      ReduxActionTypes.SELECT_MULTIPLE_WIDGETS_INIT,
+      selectAllWidgetsSaga,
+    ),
+  ]);
+}
